enhanced: Neo4COVID19 workflow enhanced by SmartGraph analysis performed via SmartGraph API call.
