+++ .gitattributes
@@ -1,1 +1,10 @@
 data/db/Neo4COVID19_PUB_rolling_neo4j.dump filter=lfs diff=lfs merge=lfs -text
+data filter=lfs diff=lfs merge=lfs -text
+data/* filter=lfs diff=lfs merge=lfs -text
+data/*.* filter=lfs diff=lfs merge=lfs -text
+data/input filter=lfs diff=lfs merge=lfs -text
+data/input/* filter=lfs diff=lfs merge=lfs -text
+data/input/*.* filter=lfs diff=lfs merge=lfs -text
+data/output filter=lfs diff=lfs merge=lfs -text
+data/output/* filter=lfs diff=lfs merge=lfs -text
+data/output/*.* filter=lfs diff=lfs merge=lfs -text
