+++ code/environment.yml
@@ -1,0 +1,20 @@
+name: n4c
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - numpy=1.19.2
+  - pandas=1.1.5
+  - pip=21.1.3
+  - python=3.6.9
+  - requests=2.25.1
+  - pip:
+    - cffi==1.14.5
+    - docker==5.0.0
+    - et-xmlfile==1.1.0
+    - neotime==1.7.4
+    - openpyxl==3.0.7
+    - py2neo==2021.1.5
+    - uuid==1.30
+    - xlrd==1.2.0
+
