--- code/n4c_env_linux.yml
@@ -1,62 +1,0 @@
-name: n4c
-channels:
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=4.5=1_gnu
-  - blas=1.0=mkl
-  - brotlipy=0.7.0=py36h27cfd23_1003
-  - ca-certificates=2021.7.5=h06a4308_1
-  - certifi=2021.5.30=py36h06a4308_0
-  - cffi=1.14.6=py36h7f8727e_0
-  - chardet=4.0.0=py36h06a4308_1003
-  - cryptography=3.4.7=py36hd23ed53_0
-  - intel-openmp=2021.3.0=h06a4308_3350
-  - libedit=3.1.20210216=h27cfd23_1
-  - libffi=3.2.1=hf484d3e_1007
-  - libgcc-ng=9.3.0=h5101ec6_17
-  - libgomp=9.3.0=h5101ec6_17
-  - libstdcxx-ng=9.3.0=hd4cf53a_17
-  - mkl=2020.2=256
-  - mkl-service=2.3.0=py36he8ac12f_0
-  - mkl_fft=1.3.0=py36h54f3939_0
-  - mkl_random=1.1.1=py36h0573a6f_0
-  - ncurses=6.2=he6710b0_1
-  - numpy=1.19.2=py36h54aff64_0
-  - numpy-base=1.19.2=py36hfa32c7d_0
-  - openssl=1.1.1k=h27cfd23_0
-  - pandas=1.1.5=py36ha9443f7_0
-  - pip=21.1.3=py36h06a4308_0
-  - pycparser=2.20=py_2
-  - pyopenssl=20.0.1=pyhd3eb1b0_1
-  - pysocks=1.7.1=py36h06a4308_0
-  - python=3.6.9=h265db76_0
-  - python-dateutil=2.8.2=pyhd3eb1b0_0
-  - pytz=2021.1=pyhd3eb1b0_0
-  - readline=7.0=h7b6447c_5
-  - setuptools=52.0.0=py36h06a4308_0
-  - six=1.16.0=pyhd3eb1b0_0
-  - sqlite=3.33.0=h62c20be_0
-  - tk=8.6.10=hbc83047_0
-  - urllib3=1.26.6=pyhd3eb1b0_1
-  - wheel=0.36.2=pyhd3eb1b0_0
-  - xz=5.2.5=h7b6447c_0
-  - zlib=1.2.11=h7b6447c_3
-  - pip:
-    - charset-normalizer==2.0.3
-    - docker==5.0.0
-    - english==2020.7.0
-    - idna==3.2
-    - monotonic==1.6
-    - neotime==1.7.4
-    - packaging==21.0
-    - pansi==2020.7.3
-    - prompt-toolkit==3.0.19
-    - py2neo==2021.1.5
-    - pygments==2.9.0
-    - pyparsing==2.4.7
-    - requests==2.26.0
-    - uuid==1.30
-    - wcwidth==0.2.5
-    - websocket-client==1.1.0
-    - xlrd==1.2.0
