--- code/n4c_env_mac.yml
@@ -1,61 +1,0 @@
-name: n4c
-channels:
-  - defaults
-  - conda-forge
-dependencies:
-  - blas=1.0=mkl
-  - brotlipy=0.7.0=py36h9ed2024_1003
-  - ca-certificates=2021.5.25=hecd8cb5_1
-  - certifi=2021.5.30=py36hecd8cb5_0
-  - chardet=4.0.0=py36hecd8cb5_1003
-  - cryptography=3.4.7=py36h2fd3fbb_0
-  - idna=2.10=pyhd3eb1b0_0
-  - intel-openmp=2021.2.0=hecd8cb5_564
-  - libcxx=10.0.0=1
-  - libedit=3.1.20210216=h9ed2024_1
-  - libffi=3.2.1=h0a44026_1007
-  - mkl=2019.4=233
-  - mkl-service=2.3.0=py36h9ed2024_0
-  - mkl_fft=1.3.0=py36ha059aab_0
-  - mkl_random=1.1.1=py36h959d312_0
-  - ncurses=6.2=h0a44026_1
-  - numpy=1.19.2=py36h456fd55_0
-  - numpy-base=1.19.2=py36hcfb5961_0
-  - openssl=1.1.1k=h9ed2024_0
-  - pandas=1.1.5=py36hb2f4e1b_0
-  - pip=21.1.3=py36hecd8cb5_0
-  - pycparser=2.20=py_2
-  - pyopenssl=20.0.1=pyhd3eb1b0_1
-  - pysocks=1.7.1=py36hecd8cb5_0
-  - python=3.6.9=h359304d_0
-  - python-dateutil=2.8.1=pyhd3eb1b0_0
-  - pytz=2021.1=pyhd3eb1b0_0
-  - readline=7.0=h1de35cc_5
-  - requests=2.25.1=pyhd3eb1b0_0
-  - setuptools=52.0.0=py36hecd8cb5_0
-  - six=1.16.0=pyhd3eb1b0_0
-  - sqlite=3.33.0=hffcf06c_0
-  - tk=8.6.10=hb0a8c7a_0
-  - urllib3=1.26.6=pyhd3eb1b0_1
-  - wheel=0.36.2=pyhd3eb1b0_0
-  - xz=5.2.5=h1de35cc_0
-  - zlib=1.2.11=h1de35cc_3
-  - pip:
-    - cffi==1.14.5
-    - docker==5.0.0
-    - english==2020.7.0
-    - et-xmlfile==1.1.0
-    - monotonic==1.6
-    - neotime==1.7.4
-    - openpyxl==3.0.7
-    - packaging==20.9
-    - pansi==2020.7.3
-    - prompt-toolkit==3.0.19
-    - py2neo==2021.1.5
-    - pygments==2.9.0
-    - pyparsing==2.4.7
-    - uuid==1.30
-    - wcwidth==0.2.5
-    - websocket-client==1.1.0
-    - xlrd==1.2.0
-
