+++ data/DISCLAIMER
@@ -1,11 +1,3 @@
-Disclaimer
-
-This software is provided on an "as is" and "as available" basis. The National Center for Advancing Translational Sciences (NCATS), a component of the National Institutes of Health, a US Federal Government Agency, expressly disclaims all warranties of any kind, whether express or implied, including, but not limited to, the implied warranties of merchantability, fitness for a particular purpose and non-infringement. NCATS makes no warranty that:
-
-a. the software will meet your requirements
-
-b. the results that may be obtained from the use of the software will be accurate or reliable
-
-
-NCATS shall have no responsibility for any damage to your computer system or loss of data that results from the download of any content, materials, information or software. NCATS reserves the right to make changes or updates to the software at any time without notice.
-
+version https://git-lfs.github.com/spec/v1
+oid sha256:a19fe7d50cfae969d02c20d0970583986c3623b3b94a2e71febf2d84feda76e5
+size 873
