+++ data/input/FIsInGene_020720_with_annotations.txt
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7b9534c3cde902ccd1be4b0cc6e2c4c01e8e803de0f8f0587bc4b9a2b57d0ac2
+size 10079353
