+++ data/input/SG_HATs_dist_3_conf_0.00.json
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:54aa19f4b000ebedcabcfb70eff7ef1a63a831d25e8c0065bbc653b466503fdd
+size 202711
