+++ data/input/SG_HATs_reverse_dist_3_conf_0.00.json
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ba6e2b7faccfbc5d9b16ac8b17e6102addfdab424f606ffb25569a99b45fd992
+size 220117
