+++ data/input/TDL_UniProt_TCRD6_rev.xlsx
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d25f45838c3775bdd43a6907f1055723ccbf341fdfe35623f2817974ce8cf716
+size 379414
