+++ data/output/FINAL_pathogen_proteins.tsv
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7f6a7dd7a2831c1726506453e5ae0704536cfafcd30303363d631f68e92674db
+size 9603
