+++ data/output/idgtdl_gene2uniprot_mapping.tsv
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f9bbe7acacbe3617589ffdf11f39152d53526306d612b2d2321df791e51c69d8
+size 422212
