+++ data/output/sg_gene2uniprot_mapping.tsv
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:baf95aa988e9355638567ff069cd3abab5009b230689891114741e0ab5282c59
+size 90525
