+++ data/output/sg_proteins_a.tsv
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d8e4ff9086dcefd2698cadb803829c07e187691d619d28fec8a73d46a1d854b2
+size 1299
