+++ data/output/sg_proteins_b.tsv
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7e728bc264171fd52b909cf742f01e1ff441aa80b157846e80daf717506cd442
+size 51801
